+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 * `[jest-runtime]` Provide `require.main` property set to module with test suite
   ([#5618](https://github.com/facebook/jest/pull/5618))
+
+### Fixes
+
+* `[babel-jest]` Remove `retainLines` argument to babel.
+  ([#5594](https://github.com/facebook/jest/pull/5594))
+
+### Chore & Maintenance
+
 * `[docs]` Add note about Node version support
   ([#5622](https://github.com/facebook/jest/pull/5622))
 * `[docs]` Update to use yarn